Standardize css layout
