+++ components/Home.module.css
@@ -1,0 +1,22 @@
+.topContainer {
+  height: 100%;
+}
+
+.container {
+  margin: auto;
+  text-align: center;
+  justify-content: center;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  gap: 1.4rem 2%;
+  width: inherit;
+}
+
+.main {
+  width: 100%;
+  max-width: 1200px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
