+++ components/Navbar.tsx
@@ -1,6 +1,6 @@
 import { Navbar, Text } from '@nextui-org/react';
 import Link from 'next/link';
-import styles from './navbar.module.css';
+import styles from './Navbar.module.css';
 
 const NitNavBar = () => {
   return (
