--- styles/Home.module.css
@@ -1,112 +1,0 @@
-.topContainer {
-  height: 100%;
-}
-
-.container {
-  margin: auto;
-  text-align: center;
-  justify-content: center;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  gap: 1.4rem 2%;
-  width: inherit;
-}
-
-.articlesContainer {
-  text-align: center;
-  justify-content: center;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  gap: 1.4rem 2%;
-  width: inherit;
-}
-
-.articlesContainerTitle {
-  margin-top: 2rem;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 1.5em;
-  color: #6e757a;
-}
-
-.main {
-  width: 100%;
-  max-width: 1200px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.articlesLink {
-  flex: 1;
-  flex-grow: 0;
-  flex-basis: 15rem;
-}
-
-.button {
-  background-color: var(--nit-color);
-  display: block;
-  height: 100%;
-  width: 100%;
-  min-height: 10rem;
-  font-weight: 400;
-}
-
-.button:hover {
-  background-color: var(--nit-color-dark);
-}
-
-@media only screen and (min-width: 1px) and (max-width: 767px) {
-  .link {
-    flex-basis: 47.5%;
-  }
-}
-
-.tagline {
-  font-size: 2.5rem;
-  font-weight: 600;
-  max-width: 1000px;
-  color: var(--nit-color);
-  margin-top: 3rem;
-}
-
-.buttonIcon {
-  padding: 1.5rem;
-  padding-left: 1rem;
-  padding-right: 0;
-  width: 3.5rem;
-}
-
-.buttonText {
-  display: block;
-  padding: 1.5rem;
-  padding-left: 0.5rem;
-  font-size: 1.8rem;
-  width: 100%;
-  hyphens: auto;
-  white-space: normal;
-}
-
-.textBox {
-  margin: 2rem auto;
-  width: 800px;
-  max-width: 95%;
-}
-
-.textBox a,
-.textBox * a {
-  color: var(--nit-color-dark);
-  text-decoration: underline;
-}
-
-.textBox p {
-  margin-bottom: 1rem;
-}
-
-.textBox ul {
-  list-style-type: disc;
-}
